hi
i didn't even come up with the questions. someone else did. such a coincidence i did most of them (but not all). there are a few that are not listed here. btw these are questions found on the homework slide in my lesson #2 about boolean and conditions. who reads these anyways? lol